backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
ultralytics==8.2.18
opencv-python-headless==4.9.0.80
websockets==12.0
pyyaml==6.0.1
PyJWT==2.8.0
